+++ src/main/java/com/tvseries/TvSeries/TvShowController.java
@@ -27,12 +27,9 @@ class TvShowController {
 
     private final TvShowRepository repository;
 
-    private RSA rsa;
-
     TvShowController(TvShowRepository repository, RSA rsa) {
 
         this.repository = repository;
-        this.rsa = rsa;
     }
 
     // Aggregate root
@@ -76,68 +73,4 @@ class TvShowController {
         repository.deleteById(id);
     }
 
-
-    @GetMapping("/auth")
-    String auth(@CookieValue("auth") String auth)
-    {
-        boolean isVerified = verifyUser(auth);
-        if (isVerified)
-            return "hello";
-        else
-            return "you are not logged in"; // возможно стоит сделать return auth()
-    }
-
-
-    @GetMapping("/register") // здесь мы даем юзеру наш публичный RSA ключ чтобы он зашифровал свои логин - пароль и направляем в register
-    String register()
-    {
-        return rsa.getPublicKey().toString();
-    }
-
-
-    @PostMapping("/register")
-    String register(@CookieValue("register") String RSAlogpass)
-    {
-        String decrypted;
-        try {
-            //System.out.println(RSAlogpass);
-            decrypted = rsa.decrypt(RSAlogpass, rsa.getPrivateKey());
-            System.out.println(decrypted);
-        } catch (IllegalBlockSizeException | InvalidKeyException | BadPaddingException | NoSuchAlgorithmException | NoSuchPaddingException e) {
-            e.printStackTrace();
-            return "";
-        } catch (Exception e) {
-            e.printStackTrace();
-            return "";
-        }
-        var regInfo = decrypted.split(":");
-        // здесь нужно добавлять пользователя в базу юзеров
-        try {
-            Algorithm algorithm = Algorithm.HMAC256(regInfo[1]); // возвращаем токен для последующей аутентификации
-            String token = JWT.create()
-                    .withIssuer("Issuer")
-                    .withClaim("user", regInfo[0])
-                    .sign(algorithm);
-            return token;
-        } catch (JWTCreationException exception){
-            return "";
-        }
-    }
-
-
-    private boolean verifyUser(String token)
-    {
-        // здесь надо выпарсить имя юзера из токена и найти его пароль в базе данных (или можно хранить в базе токен)
-        String pass = "TABURETH"; // и вставить пароль сюда
-        try {
-            Algorithm algorithm = Algorithm.HMAC256(pass);
-            JWTVerifier verifier = JWT.require(algorithm)
-                    .withIssuer("Issuer")
-                    .build();
-            DecodedJWT jwt = verifier.verify(token);
-        } catch (JWTVerificationException exception){
-            return false;
-        }
-        return true;
-    }
 }
